moved the descriptor classes to a separate file
