+++ regions/shapes/line.py
@@ -7,7 +7,7 @@ import astropy.units as u
 from astropy.wcs.utils import pixel_to_skycoord, skycoord_to_pixel
 
 from ..core import PixCoord, PixelRegion, SkyRegion, BoundingBox
-from ..core.core import ScalarPix, ScalarSky
+from ..core.attributes import ScalarPix, ScalarSky
 
 
 __all__ = ['LinePixelRegion', 'LineSkyRegion']
